+++ webwhois/settings.py
@@ -1,5 +1,5 @@
 #
-# Copyright (C) 2016-2021  CZ.NIC, z. s. p. o.
+# Copyright (C) 2016-2022  CZ.NIC, z. s. p. o.
 #
 # This file is part of FRED.
 #
@@ -15,14 +15,24 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with FRED.  If not, see <https://www.gnu.org/licenses/>.
+#
 import os
 from functools import partial
+from typing import Any, Dict
 
-from appsettings import AppSettings, FileSetting, Setting, StringSetting
+from appsettings import AppSettings, DictSetting, FileSetting, StringSetting
+from frgal import make_credentials
 
 
-def _get_logger_defalt(setting_name):
-    return getattr(WEBWHOIS_SETTINGS, setting_name)
+class LoggerOptionsSetting(DictSetting):
+    """Custom dict setting for logger options."""
+
+    def transform(self, value: Dict[str, Any]) -> Dict[str, Any]:
+        """Transform the credentials."""
+        value = super().transform(value)
+        if 'credentials' in value:
+            value['credentials'] = make_credentials(**value['credentials'])
+        return value
 
 
 class WebwhoisAppSettings(AppSettings):
@@ -32,10 +42,8 @@ class WebwhoisAppSettings(AppSettings):
     CDNSKEY_SSL_CERT = FileSetting(default=None, mode=os.R_OK)
     CORBA_NETLOC = StringSetting(default=partial(os.environ.get, 'FRED_WEBWHOIS_NETLOC', 'localhost'))
     CORBA_CONTEXT = StringSetting(default='fred')
-    LOGGER = Setting(default='pylogger.corbalogger.Logger')
-    LOGGER_CORBA_NETLOC = StringSetting(default=partial(_get_logger_defalt, 'CORBA_NETLOC'))
-    LOGGER_CORBA_CONTEXT = StringSetting(default=partial(_get_logger_defalt, 'CORBA_CONTEXT'))
-    LOGGER_CORBA_OBJECT = StringSetting(default='Logger')
+    LOGGER = StringSetting(default='grill.DummyLoggerClient')
+    LOGGER_OPTIONS = LoggerOptionsSetting(default={})
 
     class Meta:
         setting_prefix = 'WEBWHOIS_'